feat(backend-cli): Port test-users.ts to backend-cli
