+++ apps/backend-cli/src/commands/index.ts
@@ -4,3 +4,4 @@ export * from "./payment.js";
 export * from "./process.js";
 export * from "./sync.js";
 export * from "./user.js";
+export * from "./test.js";
